+++ lib/cogs/misc.py
@@ -1,7 +1,9 @@
 from discord.ext.commands import Cog
 from discord.ext.commands import CheckFailure
 from discord.ext.commands import command, has_permissions, cooldown, BucketType
-from discord import Embed
+from discord import Embed, Message, Reaction
+
+import random
 
 from ..db import db # pylint: disable=relative-beyond-top-level
 
@@ -38,6 +40,54 @@ class Misc(Cog):
 
 		for emoji in emojis:
 			await message.add_reaction(emoji)
+
+	
+	@command(name="giveaway", aliases=['startgiveaway'], brief="Creates a giveaway, that will choose a winner!")
+	@cooldown(1, 5, BucketType.user)
+	@has_permissions(administrator=True)
+	async def start_giveaway(self, ctx, *, prize: str):
+		"""Lets a Server Admin start a giveaway\n`Server Administrator` permission required."""
+		embed=Embed(title=f"{prize} giveaway!", description=f"{ctx.author.display_name} is giving away {prize}!\nReact with 🎁!")
+		embed.set_footer(text=f"{ctx.author} started this giveaway.", icon_url=ctx.author.avatar_url)
+
+		message = await ctx.send(embed=embed)
+
+		await message.add_reaction('🎁')
+
+	@command(name="stopgiveaway", aliases=['endgiveaway'], brief="Stops the giveaway, this chooses the winner!")
+	@cooldown(1, 5, BucketType.user)
+	@has_permissions(administrator=True)
+	async def stop_giveaway(self, ctx, *, message_id: Message):
+		"""Lets a Server Admin stop a giveaway, this chooses the winner at random!\n`Server Administrator` permission required."""
+
+		channel = self.bot.get_channel(message_id.channel.id)
+		message = await channel.fetch_message(message_id.id)
+
+		users = set()
+
+		for reaction in message.reactions:
+			async for user in reaction.users():
+				if not user.bot:
+					users.add(user)
+
+		entries = list()
+
+		for user in users:
+			if not user.bot:
+				entries.append(user.id)
+			else:
+				print("lol")
+
+		winner = random.choice(entries)
+
+		print(entries)
+
+		await channel.send(f"<@{winner}> won the giveaway!")
+		await channel.send(f"{ctx.author.mention}, the giveaway has been ended.", delete_after=30)
+
+		user = self.bot.get_user(winner)
+
+		await user.send(f'You won the giveaway from <#{channel.id}> in {channel.guild.name}!')
 
 	@command(name="timebomb", aliases=["tbmsg", "tb", "timebombmessage"], brief="Send a message with a timelimit on it.")
 	@cooldown(1, 4, BucketType.user)